requirements.txt
fastapi
uvicorn
python-multipart
openai
pyaudio
numpy
resampy
stream2sentence
python-dotenv
html2text
nltk
markdownify
